Cleanup and fix lint
